DRTII-1184 moving api service
