+++ client/src/main/scala/drt/client/services/handlers/ShiftsHandler.scala
@@ -1,3 +1,4 @@
+
 package drt.client.services.handlers
 
 import autowire._
