+++ server/src/main/scala/controllers/Application.scala
@@ -1,12 +1,9 @@
 
 package controllers
 
-import actors.PartitionedPortStateActor.GetStateForTerminalDateRange
 import actors._
-import actors.persistent.staffing.{GetState, UpdateShifts}
 import akka.actor._
 import akka.event.{Logging, LoggingAdapter}
-import akka.pattern._
 import akka.stream._
 import akka.util.Timeout
 import api._
@@ -16,23 +13,20 @@ import com.typesafe.config.ConfigFactory
 import controllers.application._
 import drt.http.ProdSendAndReceive
 import drt.shared.CrunchApi._
-import drt.shared.KeyCloakApi.{KeyCloakGroup, KeyCloakUser}
 import drt.shared._
 import drt.users.KeyCloakClient
 import org.joda.time.chrono.ISOChronology
 import org.slf4j.{Logger, LoggerFactory}
 import play.api.mvc._
 import play.api.{Configuration, Environment}
-import play.filters.csp.CSP
 import services.PcpArrival._
 import services._
 import services.graphstages.Crunch
 import services.metrics.Metrics
-import services.staffing.StaffTimeSlots
 import slickdb.UserTableLike
 import uk.gov.homeoffice.drt.arrivals.Arrival
-import uk.gov.homeoffice.drt.auth.Roles.{BorderForceStaff, ManageUsers, Role, StaffEdit}
-import uk.gov.homeoffice.drt.auth.{LoggedInUser, _}
+import uk.gov.homeoffice.drt.auth.Roles.{BorderForceStaff, Role}
+import uk.gov.homeoffice.drt.auth._
 import uk.gov.homeoffice.drt.ports.Terminals.Terminal
 import uk.gov.homeoffice.drt.ports.{AclFeedSource, AirportConfig, FeedSource, PortCode}
 import uk.gov.homeoffice.drt.redlist.RedListUpdates
@@ -202,127 +196,6 @@ class Application @Inject()(implicit val config: Configuration, env: Environment
     SDate(date.millisSinceEpoch - oneDayInMillis)
   }
 
-  val permissionDeniedMessage = "You do not have permission manage users"
-
-  object ApiService {
-    def apply(
-               airportConfig: AirportConfig,
-               shiftsActor: ActorRef,
-               fixedPointsActor: ActorRef,
-               staffMovementsActor: ActorRef,
-               headers: Headers,
-               session: Session
-             ): ApiService = new ApiService(airportConfig, shiftsActor, fixedPointsActor, staffMovementsActor, headers, session) {
-
-      def actorSystem: ActorSystem = system
-
-      def getLoggedInUser(): LoggedInUser = ctrl.getLoggedInUser(config, headers, session)
-
-      def forecastWeekSummary(startDay: MillisSinceEpoch,
-                              terminal: Terminal): Future[Option[ForecastPeriodWithHeadlines]] = {
-        val numberOfDays = 7
-        val (startOfForecast, endOfForecast) = startAndEndForDay(startDay, numberOfDays)
-
-        val portStateFuture = portStateActor.ask(
-          GetStateForTerminalDateRange(startOfForecast.millisSinceEpoch, endOfForecast.millisSinceEpoch, terminal)
-        )(new Timeout(30 seconds))
-
-        portStateFuture
-          .map {
-            case portState: PortState =>
-              log.info(s"Sent forecast for week beginning ${SDate(startDay).toISOString()} on $terminal")
-              val fp = services.exports.Forecast.forecastPeriod(airportConfig, terminal, startOfForecast, endOfForecast, portState)
-              val hf = services.exports.Forecast.headlineFigures(startOfForecast, numberOfDays, terminal, portState, airportConfig.queuesByTerminal(terminal).toList)
-              Option(ForecastPeriodWithHeadlines(fp, hf))
-          }
-          .recover {
-            case t =>
-              log.error(s"Failed to get PortState", t)
-              None
-          }
-      }
-
-      def updateShifts(shiftsToUpdate: Seq[StaffAssignment]): Unit = {
-        if (getLoggedInUser().roles.contains(StaffEdit)) {
-          log.info(s"Saving ${shiftsToUpdate.length} shift staff assignments")
-          shiftsActor ! UpdateShifts(shiftsToUpdate)
-        } else throw new Exception("You do not have permission to edit staffing.")
-      }
-
-      def getShiftsForMonth(month: MillisSinceEpoch, terminal: Terminal): Future[ShiftAssignments] = {
-        val shiftsFuture = shiftsActor ? GetState
-
-        shiftsFuture.collect {
-          case shifts: ShiftAssignments =>
-            log.info(s"Shifts: Retrieved shifts from actor for month starting: ${SDate(month).toISOString()}")
-            val monthInLocalTime = SDate(month, Crunch.europeLondonTimeZone)
-            StaffTimeSlots.getShiftsForMonth(shifts, monthInLocalTime)
-        }
-      }
-
-      def keyCloakClient: KeyCloakClient with ProdSendAndReceive = {
-        val token = headers.get("X-Auth-Token")
-          .getOrElse(throw new Exception("X-Auth-Token missing from headers, we need this to query the Key Cloak API."))
-        val keyCloakUrl = config.getOptional[String]("key-cloak.url")
-          .getOrElse(throw new Exception("Missing key-cloak.url config value, we need this to query the Key Cloak API"))
-        new KeyCloakClient(token, keyCloakUrl) with ProdSendAndReceive
-      }
-
-      def getKeyCloakUsers(): Future[List[KeyCloakUser]] = {
-        log.info(s"Got these roles: ${getLoggedInUser().roles}")
-        if (getLoggedInUser().roles.contains(ManageUsers)) {
-          Future(keyCloakClient.getAllUsers().toList)
-        } else throw new Exception(permissionDeniedMessage)
-      }
-
-      def getKeyCloakGroups(): Future[List[KeyCloakGroup]] = {
-        if (getLoggedInUser().roles.contains(ManageUsers)) {
-          keyCloakClient.getGroups
-        } else throw new Exception(permissionDeniedMessage)
-      }
-
-      def getKeyCloakUserGroups(userId: String): Future[Set[KeyCloakGroup]] = {
-        if (getLoggedInUser().roles.contains(ManageUsers)) {
-          keyCloakClient.getUserGroups(userId).map(_.toSet)
-        } else throw new Exception(permissionDeniedMessage)
-      }
-
-      case class KeyCloakGroups(groups: List[KeyCloakGroup])
-
-
-      def addUserToGroups(userId: String, groups: Set[String]): Future[Unit] =
-        if (getLoggedInUser().roles.contains(ManageUsers)) {
-          val futureGroupIds: Future[KeyCloakGroups] = keyCloakClient
-            .getGroups
-            .map(kcGroups => KeyCloakGroups(kcGroups.filter(g => groups.contains(g.name))))
-
-
-          futureGroupIds.map {
-            case KeyCloakGroups(gps) if gps.nonEmpty =>
-              log.info(s"Adding ${gps.map(_.name)} to $userId")
-              gps.foreach(group => {
-                val response = keyCloakClient.addUserToGroup(userId, group.id)
-                response.map(res => log.info(s"Added group and got: ${res.status}  $res")
-                )
-              })
-            case _ => log.error(s"Unable to add $userId to $groups")
-          }
-        } else throw new Exception(permissionDeniedMessage)
-
-      def removeUserFromGroups(userId: String, groups: Set[String]): Future[Unit] =
-        keyCloakClient
-          .getGroups
-          .map(kcGroups => kcGroups.filter(g => groups.contains(g.name))
-            .foreach(g => keyCloakClient.removeUserFromGroup(userId, g.id)))
-
-      override def portStateActor: ActorRef = ctrl.portStateActor
-
-      def getShowAlertModalDialog(): Boolean = config
-        .getOptional[Boolean]("feature-flags.display-modal-alert")
-        .getOrElse(false)
-
-    }
-  }
 
   def autowireApi(path: String): Action[RawBuffer] = authByRole(BorderForceStaff) {
     Action.async(parse.raw) {
@@ -332,7 +205,12 @@ class Application @Inject()(implicit val config: Configuration, env: Environment
         val b = request.body.asBytes(parse.UNLIMITED).get
 
         val router = Router.route[Api](
-          ApiService(airportConfig, ctrl.shiftsActor, ctrl.fixedPointsActor, ctrl.staffMovementsActor, request.headers, request.session))
+          new ApiService(airportConfig,
+            ctrl.shiftsActor,
+            ctrl.fixedPointsActor,
+            ctrl.staffMovementsActor,
+            request.headers,
+            request.session))
 
         router(
           autowire.Core.Request(path.split("/"), Unpickle[Map[String, ByteBuffer]].fromBytes(b.asByteBuffer))
