+++ server/src/main/scala/controllers/DrtActorSystem.scala
@@ -1,3 +1,4 @@
+
 package controllers
 
 import actors.{DrtSystemInterface, ProdDrtParameters, ProdDrtSystem}
@@ -15,13 +16,20 @@ import scala.concurrent.{ExecutionContext, ExecutionContextExecutor}
 object DrtActorSystem extends AirportConfProvider {
   val config: Configuration = new Configuration(ConfigFactory.load)
   val isTestEnvironment: Boolean = config.getOptional[String]("env").getOrElse("prod") == "test"
-  implicit val actorSystem: ActorSystem = if (isTestEnvironment) ActorSystem("DRT", PersistenceTestKitPlugin.config.withFallback(ConfigFactory.load())) else ActorSystem("DRT")
+  implicit val actorSystem: ActorSystem = if (isTestEnvironment) {
+    ActorSystem("DRT", PersistenceTestKitPlugin.config.withFallback(ConfigFactory.load()))
+  } else {
+    ActorSystem("DRT")
+  }
   implicit val mat: Materializer = Materializer.createMaterializer(actorSystem)
   implicit val ec: ExecutionContextExecutor = ExecutionContext.global
   implicit val timeout: Timeout = new Timeout(5.seconds)
   val drtSystem: DrtSystemInterface =
-    if (isTestEnvironment) drtTestSystem
-    else drtProdSystem
+    if (isTestEnvironment) {
+      drtTestSystem
+    } else {
+      drtProdSystem
+    }
 
   lazy val drtTestSystem: TestDrtSystem = TestDrtSystem(airportConfig, MockDrtParameters())
   lazy val drtProdSystem: ProdDrtSystem = ProdDrtSystem(airportConfig, ProdDrtParameters(config))
