+++ server/src/main/scala/services/ApiService.scala
@@ -1,69 +1,30 @@
+
 package services
 
-import akka.actor.{ActorRef, ActorSystem}
+import actors.PartitionedPortStateActor.GetStateForTerminalDateRange
+import actors.persistent.staffing.{GetState, UpdateShifts}
+import akka.actor._
+import akka.pattern._
+import akka.stream._
 import akka.util.Timeout
-import controllers.ShiftPersistence
+import controllers.{DrtActorSystem, ShiftPersistence}
 import drt.shared.CrunchApi._
-import drt.shared.KeyCloakApi.{KeyCloakGroup, KeyCloakUser}
-import uk.gov.homeoffice.drt.ports.Terminals.Terminal
 import drt.shared._
 import org.slf4j.{Logger, LoggerFactory}
-import play.api.mvc.{Headers, Session}
-import uk.gov.homeoffice.drt.ports.{AirportConfig, PortCode}
-import uk.gov.homeoffice.drt.redlist.RedListUpdates
+import play.api.mvc._
+import services.graphstages.Crunch
+import services.staffing.StaffTimeSlots
+import uk.gov.homeoffice.drt.auth.LoggedInUser
+import uk.gov.homeoffice.drt.auth.Roles.StaffEdit
+import uk.gov.homeoffice.drt.ports.AirportConfig
+import uk.gov.homeoffice.drt.ports.Terminals.Terminal
+import uk.gov.homeoffice.drt.time.{SDate, SDateLike}
 
-import scala.collection.immutable.Map
-import scala.concurrent.ExecutionContext.Implicits.global
-import scala.concurrent.Future
 import scala.concurrent.duration._
-import scala.io.Codec
+import scala.concurrent.{ExecutionContext, Future}
 import scala.language.postfixOps
-import scala.util.Try
 
-trait AirportToCountryLike {
-  lazy val airportInfoByIataPortCode: Map[String, AirportInfo] = {
-    val bufferedSource = scala.io.Source.fromURL(getClass.getResource("/airports.dat"))(Codec.UTF8)
-    bufferedSource.getLines().map { l =>
-      val t = Try {
-        val splitRow: Array[String] = l.split(",")
-        val sq: String => String = stripQuotes
-        AirportInfo(sq(splitRow(1)), sq(splitRow(2)), sq(splitRow(3)), sq(splitRow(4)))
-      }
-      t.getOrElse({
-        AirportInfo("failed on", l, "boo", "ya")
-      })
-    }.map(ai => (ai.code, ai)).toMap
-  }
-
-  def stripQuotes(row1: String): String = {
-    row1.substring(1, row1.length - 1)
-  }
-
-  def airportInfosByAirportCodes(codes: Set[String]): Future[Map[String, AirportInfo]] = Future {
-    val res = codes.map(code => (code, airportInfoByIataPortCode.get(code)))
-
-    val successes: Set[(String, AirportInfo)] = res collect {
-      case (code, Some(ai)) =>
-        (code, ai)
-    }
-
-    successes.toMap
-  }
-}
-
-object AirportToCountry extends AirportToCountryLike {
-  def isRedListed(portToCheck: PortCode, forDate: MillisSinceEpoch, redListUpdates: RedListUpdates): Boolean = airportInfoByIataPortCode
-    .get(portToCheck.iata)
-    .exists(ai => redListUpdates.countryCodesByName(forDate).contains(ai.country))
-}
-
-abstract class ApiService(val airportConfig: AirportConfig,
-                          val shiftsActor: ActorRef,
-                          val fixedPointsActor: ActorRef,
-                          val staffMovementsActor: ActorRef,
-                          val headers: Headers,
-                          val session: Session)
-  extends Api with ShiftPersistence {
+trait ApiServiceI extends Api with ShiftPersistence {
 
   override implicit val timeout: akka.util.Timeout = Timeout(30 seconds)
 
@@ -73,22 +34,92 @@ abstract class ApiService(val airportConfig: AirportConfig,
 
   def actorSystem: ActorSystem
 
+  def shiftsActor: ActorRef
+
   def forecastWeekSummary(startDay: MillisSinceEpoch, terminal: Terminal): Future[Option[ForecastPeriodWithHeadlines]]
 
   def getShiftsForMonth(month: MillisSinceEpoch, terminal: Terminal): Future[ShiftAssignments]
 
   def updateShifts(shiftsToUpdate: Seq[StaffAssignment]): Unit
 
-  def getKeyCloakUsers(): Future[List[KeyCloakUser]]
-
-  def getKeyCloakGroups(): Future[List[KeyCloakGroup]]
-
-  def getKeyCloakUserGroups(userId: String): Future[Set[KeyCloakGroup]]
-
-  def addUserToGroups(userId: String, groups: Set[String]): Future[Unit]
-
-  def removeUserFromGroups(userId: String, groups: Set[String]): Future[Unit]
-
   def getShowAlertModalDialog(): Boolean
 }
 
+class ApiService(airportConfig: AirportConfig,
+                 shiftsActorRef: ActorRef,
+                 fixedPointsActor: ActorRef,
+                 staffMovementsActor: ActorRef,
+                 headers: Headers, session: Session) extends ApiServiceI {
+
+  implicit val system: ActorSystem = DrtActorSystem.actorSystem
+  implicit val mat: Materializer = DrtActorSystem.mat
+  implicit val ec: ExecutionContext = DrtActorSystem.ec
+  val ctrl = DrtActorSystem.drtSystem
+  val config = DrtActorSystem.config
+  override def shiftsActor: ActorRef = shiftsActorRef
+
+  override def actorSystem: ActorSystem = DrtActorSystem.actorSystem
+
+  def getLoggedInUser(): LoggedInUser =
+    ctrl.getLoggedInUser(config, headers, session)
+
+
+  def startAndEndForDay(startDay: MillisSinceEpoch, numberOfDays: Int): (SDateLike, SDateLike) = {
+    val startOfWeekMidnight = SDate(startDay).getLocalLastMidnight
+    val endOfForecast = startOfWeekMidnight.addDays(numberOfDays)
+
+    (startOfWeekMidnight, endOfForecast)
+  }
+
+  def forecastWeekSummary(startDay: MillisSinceEpoch, terminal: Terminal):
+  Future[Option[ForecastPeriodWithHeadlines]] = {
+    val numberOfDays = 7
+    val (startOfForecast, endOfForecast) = startAndEndForDay(startDay, numberOfDays)
+
+    val portStateFuture = portStateActor.ask(
+      GetStateForTerminalDateRange(startOfForecast.millisSinceEpoch, endOfForecast.millisSinceEpoch, terminal)
+    )(new Timeout(30 seconds))
+
+    portStateFuture
+      .map {
+        case portState: PortState =>
+          log.info(s"Sent forecast for week beginning ${SDate(startDay).toISOString()} on $terminal")
+          val fp = services.exports.Forecast.forecastPeriod(airportConfig, terminal, startOfForecast, endOfForecast, portState)
+          val hf = services.exports.Forecast.headlineFigures(startOfForecast, numberOfDays, terminal, portState,
+            airportConfig.queuesByTerminal(terminal).toList)
+          Option(ForecastPeriodWithHeadlines(fp, hf))
+      }
+      .recover {
+        case t =>
+          log.error(s"Failed to get PortState", t)
+          None
+      }
+  }
+
+  def updateShifts(shiftsToUpdate: Seq[StaffAssignment]): Unit =
+    if (getLoggedInUser().roles.contains(StaffEdit)) {
+      log.info(s"Saving ${shiftsToUpdate.length} shift staff assignments")
+      shiftsActor ! UpdateShifts(shiftsToUpdate)
+    } else {
+      throw new Exception("You do not have permission to edit staffing.")
+    }
+
+
+  def getShiftsForMonth(month: MillisSinceEpoch, terminal: Terminal): Future[ShiftAssignments] = {
+    val shiftsFuture = shiftsActor ? GetState
+
+    shiftsFuture.collect {
+      case shifts: ShiftAssignments =>
+        log.info(s"Shifts: Retrieved shifts from actor for month starting: ${SDate(month).toISOString()}")
+        val monthInLocalTime = SDate(month, Crunch.europeLondonTimeZone)
+        StaffTimeSlots.getShiftsForMonth(shifts, monthInLocalTime)
+    }
+  }
+
+  def getShowAlertModalDialog(): Boolean = config
+    .getOptional[Boolean]("feature-flags.display-modal-alert")
+    .getOrElse(false)
+
+  override def portStateActor: ActorRef = ctrl.portStateActor
+
+}
